# Asset requests optimisation